Added rsyslog to pam_tacplus-client
Now we can actually see the log messages that PAM-tacplus is printing
as pamtester exercises it.

pam_tacplus-client now has rsyslogd as the entrypoint. This daemon needs
to be running before the log events so that they are actually captured
and it was unnecessary using bash in here when docker-compose exec is
used later to interact with it. rsyslogd is now used to keep the
container in the running state.