Removed apt-get install of libpam-tacplus
I couldn't get the installation in /usr/local/lib/security/ to work. It
would always use the one in /lib/security even when changing
test-pam.conf to use the fully-qualified path to the lib.

Updated the README with instructions for building such that
pam-tacplus.so is installed into /lib/security.